refactor: introduce exercise and number utils to move code from home.js to separate files
